Remove js_helper.py
